# [ENHANCEMENT] PromQL: Re-structure aggregations for clarity and performance