@bboreham 2024-03-18
Similar idea: #8602. 
(Which I don’t think I would have understood at the time. I have to try changing the code myself before I understand it)